third_party/blink/web_tests/fast/dom/node-filter-gc-expected.txt

PASS iterator.filter.foo is undefined.
PASS iterator.filter.foo is 42
PASS iterator.filter.foo is 42
PASS walker.filter.foo is undefined.
PASS walker.filter.foo is 42
PASS walker.filter.foo is 42
PASS successfullyParsed is true

TEST COMPLETE
